Commit after db merge
I was able to load the mean rent into the geojson file with the neighborhood polygons through mongoDB (file nyc_neighborhoods.json). Everyone should be able to access it and load it into their mongo now!
I loaded it using the node insert to Mongo files from Dartanion, the original files are in folder 19 of the class repo.